New grid library
Adds a new grid library to standardize/share grid operations across
completed challenges. This includes types for points and lines, as well
as an `isAdjacent(point: Point, line: Line, offset = 1)` function to
check whether a given point is adjacent to a given line and offset. This
will be useful for completed challenges that utilize grid operations,
such as day 3.